requirements.txt
pydantic
fastapi
uvicorn
python-decouple
jsonschema[format]
fastapi-mail
pytest-asyncio